requirements.txt
numpy>=1.22
pyyaml>=5.0
opencv-python
glob